+++ src/apps/backend/modules/authentication/internals/otp/otp_util.py
@@ -2,9 +2,9 @@ import secrets
 import string
 from typing import Any
 
-from modules.authentication.internals.otp.internal.store.otp_model import OTPModel
-from modules.config.config_service import ConfigService
+from modules.authentication.internals.otp.store.otp_model import OTPModel
 from modules.authentication.types import OTP
+from modules.config.config_service import ConfigService
 
 
 class OTPUtil:
