+++ src/apps/backend/modules/authentication/internals/otp/otp_writer.py
@@ -4,9 +4,9 @@ from pymongo import ReturnDocument
 
 from modules.account.types import PhoneNumber
 from modules.authentication.errors import OTPExpiredError, OTPIncorrectError
-from modules.authentication.internals.otp.internal.otp_util import OTPUtil
-from modules.authentication.internals.otp.internal.store.otp_model import OTPModel
-from modules.authentication.internals.otp.internal.store.otp_repository import OTPRepository
+from modules.authentication.internals.otp.otp_util import OTPUtil
+from modules.authentication.internals.otp.store.otp_model import OTPModel
+from modules.authentication.internals.otp.store.otp_repository import OTPRepository
 from modules.authentication.types import CreateOTPParams, OTP, OTPStatus, VerifyOTPParams
